+++ firestore-counter/functions/src/controller.ts
@@ -19,6 +19,7 @@ import { Slice, WorkerStats, queryRange } from "./common";
 import { Planner } from "./planner";
 import { Aggregator } from "./aggregator";
 import { logger } from "firebase-functions";
+import { FieldValue } from "firebase-admin/firestore";
 
 export interface WorkerShardingInfo {
   slice: Slice; // shard range a single worker is responsible for
@@ -70,7 +71,7 @@ export class ShardedCounterController {
     limit: number,
     timeoutMillis: number
   ) {
-    return new Promise((resolve, reject) => {
+    return new Promise<void>((resolve, reject) => {
       let aggrPromise: Promise<ControllerStatus> = null;
       let controllerData: ControllerData = EMPTY_CONTROLLER_DATA;
       let rounds = 0;
@@ -93,7 +94,7 @@ export class ShardedCounterController {
         limit
       ).onSnapshot(async (snap) => {
         if (snap.docs.length == limit) return;
-        if (controllerData.workers.length > 0) {
+        if (controllerData.workers && controllerData.workers.length > 0) {
           skippedRoundsDueToWorkers++;
           return;
         }
@@ -149,9 +150,21 @@ export class ShardedCounterController {
           );
           throw Error("Failed to read controller doc.");
         }
-        const controllerData: ControllerData = controllerDoc.exists
-          ? controllerDoc.data()
-          : EMPTY_CONTROLLER_DATA;
+        let controllerData: ControllerData;
+        if (controllerDoc.exists) {
+          controllerData = controllerDoc.data();
+        } else {
+          // If we arrive here, then it is the very first run of the function
+          // and the controllerDoc document has not been created, yet.
+          //
+          // We expect the controllerDoc document to have a certain structure.
+          // Therefore, We create an empty document here and exit immediately,
+          // mainly, because
+          // (a) its the first run and aggregrations will not be necessary
+          // (b) writes in transactions have to happen after all reads.
+          await t.set(this.controllerDocRef, EMPTY_CONTROLLER_DATA);
+          return ControllerStatus.SUCCESS;
+        }
         if (controllerData.workers.length > 0)
           return ControllerStatus.WORKERS_RUNNING;
 
@@ -203,7 +216,7 @@ export class ShardedCounterController {
         }
         t.set(
           this.controllerDocRef,
-          { timestamp: firestore.FieldValue.serverTimestamp() },
+          { timestamp: FieldValue.serverTimestamp() },
           { merge: true }
         );
         logger.log("Aggregated " + plans.length + " counters.");
@@ -300,13 +313,13 @@ export class ShardedCounterController {
             ),
             {
               slice: slice,
-              timestamp: firestore.FieldValue.serverTimestamp(),
+              timestamp: FieldValue.serverTimestamp(),
             }
           );
         });
         t.set(this.controllerDocRef, {
           workers: slices,
-          timestamp: firestore.FieldValue.serverTimestamp(),
+          timestamp: FieldValue.serverTimestamp(),
         });
       } else {
         // Check workers that haven't updated stats for over 90s - they most likely failed.
@@ -315,7 +328,7 @@ export class ShardedCounterController {
           if (timestamp / 1000 - snap.updateTime.seconds > 90) {
             t.set(
               snap.ref,
-              { timestamp: firestore.FieldValue.serverTimestamp() },
+              { timestamp: FieldValue.serverTimestamp() },
               { merge: true }
             );
             failures++;
@@ -324,7 +337,7 @@ export class ShardedCounterController {
         logger.log("Detected " + failures + " failed workers.");
         t.set(
           this.controllerDocRef,
-          { timestamp: firestore.FieldValue.serverTimestamp() },
+          { timestamp: FieldValue.serverTimestamp() },
           { merge: true }
         );
       }